Clean up default.qubit (#477)
* current changes

* more

* more

* tests passing

* revert plugin guide

* cleaning up

* linting

* added gate data file

* fix tests

* fixes

* test fixes

* Add some default values to the QubitDevice capabilities dictionary (#478)
