+++ tests/qnodes/test_qnode_qubit.py
@@ -90,7 +90,7 @@ class TestExpectationJacobian:
         """Tests that the correct gradient is computed for qnodes which
         use the same parameter in multiple gates."""
 
-        from pennylane.plugins.default_qubit import Rotx as Rx, Roty as Ry, Rotz as Rz
+        from gate_data import Rotx as Rx, Roty as Ry, Rotz as Rz
 
         def expZ(state):
             return np.abs(state[0]) ** 2 - np.abs(state[1]) ** 2
